+++ changelog.md
@@ -1,5 +1,9 @@
 Versions history:
 ----
+- 2019-01-10. Version 5.0.60 - Development Test 38: Update to Zotero 5.0.60.
+  
+- 2018-12-19. Version 5.0.59 - Development Test 37: Update to Zotero 5.0.59.
+  
 - 2018-11-12. Version 5.0.58 - Development Test 36: Update to Zotero 5.0.58.
   
 - 2018-10-24. Version 5.0.57 - Development Test 35: Update to Zotero 5.0.57.